feat: draw unit range

+++ src/game_unit.rs
@@ -5,16 +5,16 @@ pub struct GameUnit {
     pub level: Age,
     pub r#type: UnitType,
     pub hp: i32,
-    pub range: i32,
+    pub range: f32,
 }
 
 impl GameUnit {
     pub fn new(level: Age, typ: UnitType) -> Self {
         let (hp, range) = match typ {
-            UnitType::Meele => (40, 1),
-            UnitType::Ranged => (20, 2),
-            UnitType::Tank => (80, 1),
-            UnitType::Super => (160, 3),
+            UnitType::Meele => (40, 1.),
+            UnitType::Ranged => (20, 2.),
+            UnitType::Tank => (80, 1.),
+            UnitType::Super => (160, 3.),
         };
         Self {
             level: level,
